.ipynb_checkpoints/requirements-checkpoint.txt
Flask==2.0.2
line-bot-sdk==2.0.1
pandas==1.3.2
numpy==1.19.5
firebase-admin==5.1.0